renaming ++ adding error catcher
